Upgrade to React 19 + Vite, replace webpack/babel toolchain
- React 16 -> 19 (createRoot replaces ReactDOM.render in src/main.jsx,
  wrapped in StrictMode)
- Webpack 4 / babel-loader -> Vite 6 + @vitejs/plugin-react
- redux ^4 -> ^5, react-redux ^5 -> ^9 (connect HOC still used by App)
- react-addons-css-transition-group -> react-transition-group with
  CSSTransition + nodeRef (required by React 19; findDOMNode is gone)
- Move public/index.html to root index.html with Vite-style module
  script tag; fix malformed viewport meta
- Rename JSX-bearing .js files to .jsx; src/index.js -> src/main.jsx
- Drop webpack.config.js, .babelrc, stale package-lock.json
- package.json: "type": "module", new scripts (dev/build/preview)

No build run in this commit per request; package-lock.json will be
regenerated on the next 'npm install'.
