--- .babelrc
@@ -1,5 +1,0 @@
-{
-  "presets": ["@babel/env", "@babel/preset-react"],
-
-  "plugins": ["@babel/plugin-proposal-class-properties"]
-}
