+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "react-connect4",
-  "version": "0.1.0",
-  "description": "A remaking of the classic Connect 4 game, using React, Webpack, Redux and Babel",
-  "main": "index.js",
+  "version": "0.2.0",
+  "private": true,
+  "description": "A remaking of the classic Connect 4 game, using React, Vite and Redux",
+  "type": "module",
   "scripts": {
-    "start": "webpack-dev-server --mode-development",
-    "build": "node_modules/.bin/webpack --config webpack.prod.js --mode production",
-    "dist": "webpack",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "repository": {
     "type": "git",
@@ -16,35 +16,19 @@
   "author": "Vic",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/victoriousj/react-min-via/issues"
+    "url": "https://github.com/victoriousj/react-connect4/issues"
   },
   "homepage": "https://github.com/victoriousj/react-connect4#readme",
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
   "dependencies": {
-    "react": "^16.5.2",
-    "react-addons-css-transition-group": "^15.6.2",
-    "react-dom": "^16.5.2",
-    "react-redux": "^5.0.7",
-    "redux": "^4.0.1"
+    "prop-types": "^15.8.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-redux": "^9.2.0",
+    "react-transition-group": "^4.4.5",
+    "redux": "^5.0.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.2",
-    "@babel/core": "^7.1.2",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "@babel/preset-env": "^7.1.0",
-    "@babel/preset-react": "^7.0.0",
-    "babel-loader": "^8.0.4",
-    "css-loader": "^1.0.0",
-    "file-loader": "^2.0.0",
-    "prop-types": "^15.6.2",
-    "style-loader": "^0.23.1",
-    "webpack": "^4.21.0",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.14"
+    "@vitejs/plugin-react": "^4.3.4",
+    "vite": "^6.0.0"
   }
 }
