--- public/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <link rel="icon" type="image/png" sizes="32x32" href="/favicon.ico">
-        <meta name="viewport" content="width=device-width", initial-scale="1" shrink-to-fit="no">
-        <title>Connect4</title>
-    </head>
-
-    <body>
-        <div id="root">
-            <noscript>
-                You need JavaScript in order to run this app.
-            </noscript>
-            <script src="../dist/bundle.js"></script>
-        </div>
-    </body>
-</html>
