--- src/index.js
@@ -1,16 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { createStore } from "redux";
-import { Provider } from "react-redux";
-
-import "./App.css";
-import App from "./App";
-import reducer from "./reducers/interactions";
-
-ReactDOM.render(
-  <Provider store={createStore(reducer)}>
-    <App />
-  </Provider>,
-
-  document.getElementById("root")
-);
