+++ src/main.jsx
@@ -1,0 +1,18 @@
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import { createStore } from 'redux';
+import { Provider } from 'react-redux';
+
+import './App.css';
+import App from './App';
+import reducer from './reducers/interactions';
+
+const store = createStore(reducer);
+
+createRoot(document.getElementById('root')).render(
+  <StrictMode>
+    <Provider store={store}>
+      <App />
+    </Provider>
+  </StrictMode>
+);
